requirements.txt
cmake==4.1.0
filelock==3.19.1
fsspec==2025.9.0
gradio==3.50.2
jinja2==3.1.6
markupsafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
pillow==10.4.0
pip==25.2
pymupdf==1.24.10
setuptools==65.5.0
sympy==1.14.0
torch==2.8.0
typing-extensions==4.15.0
uv==0.8.12
easyocr==1.7.1
